requirements.txt
crewai
langchain_ibm
python-dotenv